feat: '프론트 게시판 뷰 작성'
기존의 파일들 폴더로 정리, 프론트에서 게시판 뷰 작성
