+++ frontend/src/App.vue
@@ -9,8 +9,8 @@
 </template>
 
 <script setup>
-import HeaderComponent from '@/components/HeaderComponent.vue'
-import FooterComponent from '@/components/FooterComponent.vue'
+import HeaderComponent from '@/components/section/HeaderComponent.vue'
+import FooterComponent from '@/components/section/FooterComponent.vue'
 </script>
 
 <style>
