--- frontend/src/components/FooterComponent.vue
@@ -1,15 +1,0 @@
-<template>
-  <footer class="py-4 bg-secondary text-white text-center mt-auto">
-    <div class="container">
-      <small>&copy; 2025 냠냠코치. 모든 권리 보유.</small>
-    </div>
-    
-  </footer>
-</template>
-
-<script setup>
-</script>
-
-<style scoped>
-/* Footer가 항상 페이지 하단에 붙게 만들고 싶으면 App.vue 전체 레이아웃을 flex column으로 조정하면 돼 */
-</style>
