+++ frontend/src/components/mainpage/TodayDiet.vue
@@ -21,7 +21,9 @@
             <div class="fw-bold">{{ course.course_name }}</div>
           </div>
         </div>
-        <div v-if="courses.length === 0" class="text-muted mt-3">오늘의 식단 정보가 없습니다.</div>
+        <div v-if="courses && courses.length === 0" class="text-muted mt-3">
+          오늘의 식단 정보가 없습니다.
+        </div>
       </div>
     </div>
 
@@ -379,7 +381,6 @@ const fetchTodayCourses = async () => {
   try {
     const response = await axios.get('/api/courses/today')
     courses.value = response.data.courses
-    console.log('courses', courses.value)
     courses.value.forEach((course) => {
       selectedCourses[course.type] = false
       course.dishes.forEach((dish) => {
@@ -466,251 +467,6 @@ watch(
   },
 )
 </script>
-
-<!-- <script>
-import axios from 'axios'
-
-export default {
-  data() {
-    return {
-      showModal: false,
-      selectedCourse: 'A',
-      courses: [],
-      selectedDishes: {},
-      selectedCourses: {},
-      selectAllMeals: false,
-      totalCalories: 750,
-      _chartRenderKey: 0, // 강제 리렌더링을 위한 키
-      colors: [
-        'rgba(255, 99, 132, 0.7)',
-        'rgba(54, 162, 235, 0.7)',
-        'rgba(255, 206, 86, 0.7)',
-        'rgba(75, 192, 192, 0.7)',
-        'rgba(153, 102, 255, 0.7)',
-      ],
-    }
-  },
-  computed: {
-    getCurrentCourseName() {
-      const selectedCourse = this.courses.find((c) => c.type === this.selectedCourse)
-      return selectedCourse ? selectedCourse.course_name : ''
-    },
-    currentCourseDishes() {
-      const selectedCourse = this.courses.find((c) => c.type === this.selectedCourse)
-      return selectedCourse ? selectedCourse.dishes : []
-    },
-    selectedDishesCalories() {
-      // 깊은 복사 없이 배열 생성 (최적화)
-      const dishesData = []
-
-      for (const course of this.courses) {
-        for (const dish of course.dishes) {
-          if (this.selectedDishes[dish.id]) {
-            dishesData.push({
-              name: dish.name,
-              calories: Number(dish.calorie || 0),
-            })
-          }
-        }
-      }
-
-      return dishesData
-    },
-    // 선택한 음식의 모든 영양소 데이터를 음식별로 반환
-    selectedDishesNutrients() {
-      const dishesData = []
-
-      for (const course of this.courses) {
-        for (const dish of course.dishes) {
-          if (this.selectedDishes[dish.id]) {
-            dishesData.push({
-              name: dish.name,
-              calories: Number(dish.calorie || 0),
-              protein: Number(dish.protein || 0),
-              carbohydrate: Number(dish.carbohydrate || 0),
-              fat: Number(dish.fat || 0),
-              sugar: Number(dish.sugarG || 0),
-            })
-          }
-        }
-      }
-
-      return dishesData
-    },
-    // 모든 영양소의 총합
-    totalNutrientValues() {
-      const totals = {
-        calories: 0,
-        protein: 0,
-        carbohydrate: 0,
-        fat: 0,
-        sugar: 0,
-      }
-
-      for (const dish of this.selectedDishesNutrients) {
-        totals.calories += dish.calories
-        totals.protein += dish.protein
-        totals.carbohydrate += dish.carbohydrate
-        totals.fat += dish.fat
-        totals.sugar += dish.sugar
-      }
-
-      return totals
-    },
-    // 각 영양소의 최대값 계산 (그래프 스케일링용)
-    maxValues() {
-      const max = {
-        calories: 0,
-        protein: 0,
-        carbohydrate: 0,
-        fat: 0,
-        sugar: 0,
-      }
-
-      for (const dish of this.selectedDishesNutrients) {
-        max.calories = Math.max(max.calories, dish.calories)
-        max.protein = Math.max(max.protein, dish.protein)
-        max.carbohydrate = Math.max(max.carbohydrate, dish.carbohydrate)
-        max.fat = Math.max(max.fat, dish.fat)
-        max.sugar = Math.max(max.sugar, dish.sugar)
-      }
-
-      // 최소값 설정 (0으로 나누기 방지)
-      max.calories = max.calories || 100
-      max.protein = max.protein || 20
-      max.carbohydrate = max.carbohydrate || 20
-      max.fat = max.fat || 20
-      max.sugar = max.sugar || 20
-
-      return max
-    },
-  },
-  watch: {
-    selectAllMeals(newValue) {
-      this.courses.forEach((course) => {
-        this.selectedCourses[course.type] = newValue
-        course.dishes.forEach((dish) => {
-          this.selectedDishes[dish.id] = newValue
-        })
-      })
-    },
-    selectedDishes: {
-      deep: true,
-      handler() {
-        this.courses.forEach((course) => {
-          const allSelected = course.dishes.every((dish) => this.selectedDishes[dish.id])
-          const noneSelected = course.dishes.every((dish) => !this.selectedDishes[dish.id])
-
-          if (allSelected) {
-            this.selectedCourses[course.type] = true
-          } else if (noneSelected) {
-            this.selectedCourses[course.type] = false
-          }
-
-          // 전체 선택 상태 갱신
-          const allCoursesSelected = this.courses.every((c) => this.selectedCourses[c.type])
-          if (allCoursesSelected !== this.selectAllMeals) {
-            this.selectAllMeals = allCoursesSelected
-          }
-        })
-
-        // 그래프 강제 업데이트
-        this._chartRenderKey++
-      },
-    },
-    selectedCourse(newVal) {
-      // 코스 선택 시 선택된 음식과 차트 초기화
-      this.courses.forEach((course) => {
-        course.dishes.forEach((dish) => {
-          this.selectedDishes[dish.id] = false
-        })
-        this.selectedCourses[course.type] = false
-      })
-      this.selectAllMeals = false
-      this._chartRenderKey++
-    },
-    showModal(val) {
-      if (val) {
-        // 모달이 열리면 강제 리렌더링
-        this._chartRenderKey++
-      }
-    },
-  },
-  methods: {
-    fetchTodayCourses() {
-      const response = axios.get('/api/courses/today')
-      console.log(response)
-      this.courses = response.data.courses
-
-      // 초기화: 모든 요리에 대한 선택 상태를 설정
-      this.courses.forEach((course) => {
-        this.selectedCourses[course.type] = false
-        course.dishes.forEach((dish) => {
-          this.selectedDishes[dish.id] = false
-        })
-      })
-
-      console.log('오늘의 코스 데이터:', this.courses)
-      // } catch (error) {
-      //   console.error('오늘의 코스 조회 실패:', error)
-      // }
-    },
-    getDishImage(dishName) {
-      // 실제 이미지 URL로 대체 필요
-      return `https://via.placeholder.com/120x80?text=${encodeURIComponent(dishName)}`
-    },
-    calculateTotalCalories(course) {
-      return course.dishes.reduce((total, dish) => total + dish.calorie, 0)
-    },
-    saveMealRecord() {
-      // 선택한 음식 정보 수집
-      const selectedMeals = []
-
-      this.courses.forEach((course) => {
-        course.dishes.forEach((dish) => {
-          if (this.selectedDishes[dish.id]) {
-            selectedMeals.push({
-              dishId: dish.id,
-              dishName: dish.name,
-              calories: dish.calorie,
-              protein: dish.protein,
-              carbohydrate: dish.carbohydrate,
-              fat: dish.fat,
-              sugar: dish.sugarG,
-              courseType: course.type,
-              courseName: course.course_name,
-            })
-          }
-        })
-      })
-
-      console.log('저장된 식사 기록:', selectedMeals)
-
-      // 여기에 API 호출 코드 추가 가능
-      // axios.post('/api/meal-records', { meals: selectedMeals });
-
-      this.showModal = false
-    },
-    increaseTotalCalories() {
-      this.totalCalories += 50
-    },
-    decreaseTotalCalories() {
-      if (this.totalCalories > 50) {
-        this.totalCalories -= 50
-      }
-    },
-    toggleCourseDishes(course) {
-      const isSelected = this.selectedCourses[course.type]
-      course.dishes.forEach((dish) => {
-        this.selectedDishes[dish.id] = isSelected
-      })
-    },
-  },
-  mounted() {
-    this.fetchTodayCourses()
-  },
-}
-</script> -->
 
 <style scoped>
 .dish-item {
