+++ frontend/src/router/index.js
@@ -5,55 +5,55 @@ const routes = [
   {
     path: '/',
     name: 'HomeView',
-    component: () => import('../views/HomeView.vue'),
+    component: () => import('../views/home/HomeView.vue'),
     meta: { title: '홈' },
   },
   {
     path: '/my',
     name: 'MyView',
-    component: () => import('../views/MyView.vue'),
+    component: () => import('../views/user/MyView.vue'),
     meta: { title: '마이페이지', requiresAuth: true },
   },
   {
     path: '/my/info',
     name: 'MyInfoView',
-    component: () => import('../views/MyInfoView.vue'),
+    component: () => import('../views/user/MyInfoView.vue'),
     meta: { title: '내 정보 | 마이페이지', requiresAuth: true },
   },
   {
-    path: '/community',
-    name: 'CommunityView',
-    component: () => import('../views/CommunityView.vue'),
+    path: '/board',
+    name: 'BoardView',
+    component: () => import('../views/board/BoardView.vue'),
     meta: { title: '커뮤니티' },
   },
   {
     path: '/survey',
     name: 'SurveyView',
-    component: () => import('../views/SurveyView.vue'),
+    component: () => import('../views/user/SurveyView.vue'),
     meta: { title: '설문조사' },
   },
   {
     path: '/login',
     name: 'LoginView',
-    component: () => import('../views/LoginView.vue'),
+    component: () => import('../views/user/LoginView.vue'),
     meta: { title: '로그인' },
   },
   {
     path: '/signup',
     name: 'SignupView',
-    component: () => import('../views/SignupView.vue'),
+    component: () => import('../views/user/SignupView.vue'),
     meta: { title: '회원가입' },
   },
   {
     path: '/features',
     name: 'FeaturesView',
-    component: () => import('../views/FeaturesView.vue'),
+    component: () => import('../views/others/FeaturesView.vue'),
     meta: { title: '기능소개' },
   },
   {
     path: '/error',
     name: 'notFound',
-    component: () => import('../views/ErrorView.vue'),
+    component: () => import('../views/error/ErrorView.vue'),
     meta: { title: '에러' },
   },
   {
